refactor(exec): extract process launcher helpers
