+++ docs/Tasks/003-refactor-file-size/tasks.md
@@ -21,7 +21,7 @@ description: "Refactor oversized source files: max 500 LOC, aim <300, avoid part
 - [ ] T013 Refactor `CodexSessionLocator` to remove `partial` (replace `[GeneratedRegex]`; extract helpers).
 - [ ] T014 Refactor `CodexStructuredOutputExtensions` (split into internal helpers; keep public extensions).
 - [x] T015 Refactor `ResilientCodexAppServerClient` (extract adapter types so the file is <500).
-- [ ] T016 Refactor `CodexProcessLauncher` (extract IO/diagnostics helpers; keep tested internals).
+- [x] T016 Refactor `CodexProcessLauncher` (extract IO/diagnostics helpers; keep tested internals).
 - [ ] T017 Ensure all `src/` files now meet the 500-line max (aim <300 where reasonable).
 
 ## Phase 2 — Validation
